进行代码的瘦身和优化工作
