+++ build.gradle
@@ -5,7 +5,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.2'
+        classpath 'com.android.tools.build:gradle:2.2.3'
         classpath "net.rdrei.android.buildtimetracker:gradle-plugin:0.5.+"
         classpath 'com.getkeepsafe.dexcount:dexcount-gradle-plugin:0.5.2'
