+++ share/src/main/java/com/liulishuo/share/LoginManager.java
@@ -1,14 +1,14 @@
 package com.liulishuo.share;
 
-import com.liulishuo.share.qq.SL_QQLoginActivity;
-import com.liulishuo.share.type.LoginType;
-import com.liulishuo.share.weibo.SL_WeiBoLoginActivity;
-import com.liulishuo.share.weixin.WeiXinLoginManager;
-
 import android.app.Activity;
 import android.content.Intent;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+
+import com.liulishuo.share.activity.SL_QQHandlerActivity;
+import com.liulishuo.share.activity.SL_WeiBoHandlerActivity;
+import com.liulishuo.share.activity.SL_WeiXinHandlerActivity;
+import com.liulishuo.share.type.LoginType;
 
 import static com.liulishuo.share.type.LoginType.QQ;
 import static com.liulishuo.share.type.LoginType.WEIBO;
@@ -20,16 +20,18 @@ import static com.liulishuo.share.type.LoginType.WEIXIN;
  */
 public class LoginManager {
 
-    public static
     @Nullable
-    LoginListener listener;
+    public static LoginListener listener;
 
     public static void login(@NonNull Activity activity, @LoginType String type, @Nullable LoginListener listener) {
         LoginManager.listener = listener;
         switch (type) {
             case QQ:
                 if (ShareBlock.isQQInstalled(activity)) {
-                    activity.startActivity(new Intent(activity, SL_QQLoginActivity.class));
+                    activity.startActivity(
+                            new Intent(activity, SL_QQHandlerActivity.class)
+                                    .putExtra(ShareBlock.KEY_IS_LOGIN_TYPE, true)
+                    );
                     activity.overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out);
                 } else {
                     if (listener != null) {
@@ -38,12 +40,15 @@ public class LoginManager {
                 }
                 break;
             case WEIBO:
-                activity.startActivity(new Intent(activity, SL_WeiBoLoginActivity.class));
+                activity.startActivity(
+                        new Intent(activity, SL_WeiBoHandlerActivity.class)
+                                .putExtra(ShareBlock.KEY_IS_LOGIN_TYPE, true)
+                );
                 activity.overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out);
                 break;
             case WEIXIN:
                 if (ShareBlock.isWeiXinInstalled(activity)) {
-                    new WeiXinLoginManager().login(activity.getApplicationContext());
+                    SL_WeiXinHandlerActivity.login(activity.getApplicationContext());
                 } else {
                     if (listener != null) {
                         listener.onError("未安装微信");
