--- share/src/main/java/com/liulishuo/share/weixin/WeiXinHandlerActivity.java
@@ -1,65 +1,0 @@
-package com.liulishuo.share.weixin;
-
-import com.liulishuo.share.LoginManager;
-import com.liulishuo.share.ShareBlock;
-import com.liulishuo.share.ShareManager;
-import com.tencent.mm.sdk.modelbase.BaseReq;
-import com.tencent.mm.sdk.modelbase.BaseResp;
-import com.tencent.mm.sdk.modelmsg.SendAuth;
-import com.tencent.mm.sdk.openapi.IWXAPI;
-import com.tencent.mm.sdk.openapi.IWXAPIEventHandler;
-import com.tencent.mm.sdk.openapi.WXAPIFactory;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-
-/**
- * Created by echo on 5/19/15.
- * 用来处理微信登录、微信分享的activity。这里真不知道微信非要个activity干嘛，愚蠢的设计!
- * 参考文档:https://open.weixin.qq.com/cgi-bin/showdocument?action=dir_list&t=resource/res_list&verify=1&id=open1419317853&lang=zh_CN
- */
-public abstract class WeiXinHandlerActivity extends Activity implements IWXAPIEventHandler {
-
-    /**
-     * BaseResp的getType函数获得的返回值，1:第三方授权， 2:分享
-     */
-    private static final int TYPE_LOGIN = 1;
-
-    private IWXAPI api;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        api = WXAPIFactory.createWXAPI(this, ShareBlock.Config.weiXinAppId, true);
-        api.handleIntent(getIntent(), this);
-        finish();
-    }
-
-    @Override
-    protected void onNewIntent(Intent intent) {
-        super.onNewIntent(intent);
-        if (api != null) {
-            api.handleIntent(getIntent(), this);
-        }
-        finish();
-    }
-
-    @Override
-    public void onReq(BaseReq baseReq) {
-        finish();
-    }
-
-    @Override
-    public void onResp(BaseResp resp) {
-        if (resp != null) {
-            if (resp instanceof SendAuth.Resp && resp.getType() == TYPE_LOGIN) {
-                WeiXinLoginManager.parseLoginResp(this, (SendAuth.Resp) resp, LoginManager.listener);
-            } else {
-                WeiXinShareManager.parseShareResp(resp, ShareManager.listener);
-            }
-        }
-        finish();
-    }
-
-}
